+++ src/components/profile/reducers.js
@@ -1,2 +1,30 @@
+export const USER_AUTH = 'USER_AUTH';
+export const LOGOUT = 'LOGOUT';
+export const CHECKED_AUTH = 'CHECKED_AUTH';
+export const USER_UPDATE = 'USER_UPDATE';
 
+export const getUser = state => state.user;
+export const getCheckedAuth = state => state.checkedAuth;
 export const getWatchlist = state => state.user.watchlist;
+
+export function user(state = null, { type, payload }) {
+  switch(type) {
+    case USER_AUTH:
+      return payload;
+    case USER_UPDATE:
+      return payload;
+    case LOGOUT:
+      return null;
+    default:
+      return state;
+  }
+}
+
+export function checkedAuth(state = false, { type }) {
+  switch(type) {
+    case CHECKED_AUTH:
+      return true;
+    default:
+      return state;
+  }
+}
